requirements.txt
streamlit
numpy
pillow
keras
tensorflow
